Fix multiple definition

+++ include/fty_common_rest_tokens.h
@@ -90,7 +90,7 @@ private:
     std::multimap<long int, std::string> revoked_queue;
     void clean_revoked();
     void regen_keys (long int expires_in);
-    static const uint16_t MESSAGE_LEN;
+    static const uint32_t MESSAGE_LEN;
 public:
     //! Singleton get_instance method
     static tokens* get_instance();
